Create js file and move function on that file
